Update time input
