+++ src/components/form/DatetimeRange.vue
@@ -8,11 +8,11 @@
       <div class="datetime-container row justify-evenly items-center">
         <ni-date-input :value="value.startDate" @input="update($event, 'startDate')" class="date-item"
           @blur="blurHandler" :disable="disable" @error="childErrors.startDate = $event" />
-        <ni-select :value="value.startHour" @input="update($event, 'startHour')" class="time-item"
-          @blur="blurHandler" :options="hoursOptions" :disable="disable" />
+        <ni-time-input :value="value.startHour" @input="update($event, 'startHour')" class="time-item"
+          @blur="blurHandler" :disable="disable" />
         <p class="delimiter">-</p>
-        <ni-select :value="value.endHour" @input="update($event, 'endHour')" class="time-item"
-          @blur="blurHandler" :options="endHourOptions" :disable="disable" />
+        <ni-time-input :value="value.endHour" @input="update($event, 'endHour')" class="time-item"
+          @blur="blurHandler" :disable="disable" />
         <ni-date-input :value="value.endDate" @input="update($event, 'endDate')" class="date-item"
           @blur="blurHandler" :min="value.startDate" :disable="disable || disableEndDate" />
       </div>
@@ -23,14 +23,13 @@
 <script>
 import { required } from 'vuelidate/lib/validators';
 import DateInput from './DateInput';
-import Select from './Select';
-import { PLANNING_VIEW_START_HOUR, PLANNING_VIEW_END_HOUR } from '../../data/constants.js';
+import TimeInput from './TimeInput';
 import { minDate, validHour } from '../../helpers/vuelidateCustomVal.js';
 
 export default {
   components: {
     'ni-date-input': DateInput,
-    'ni-select': Select,
+    'ni-time-input': TimeInput,
   },
   props: {
     caption: { type: String, default: '' },
@@ -61,21 +60,6 @@ export default {
     }
   },
   computed: {
-    hoursOptions () {
-      const range = this.$moment.range(
-        this.$moment().hours(PLANNING_VIEW_START_HOUR).minutes(0),
-        this.$moment().hours(PLANNING_VIEW_END_HOUR).minutes(0)
-      );
-      const hours = Array.from(range.by('hours'));
-      const selectOptions = [];
-      hours.map((hour) => {
-        selectOptions.push({ label: hour.format('HH:mm'), value: hour.format('HH:mm') });
-        if (hour.format('HH') !== `${PLANNING_VIEW_END_HOUR}`) {
-          selectOptions.push({ label: hour.minutes(30).format('HH:mm'), value: hour.minutes(30).format('HH:mm') });
-        }
-      });
-      return selectOptions;
-    },
     hasError () {
       if (!this.blur) return false;
 
@@ -90,11 +74,6 @@ export default {
       const endDatetime = this.$moment(this.value.endDate).hours(endTime[0]).minutes(endTime[1]);
 
       return startDatetime.isAfter(endDatetime);
-    },
-    endHourOptions () {
-      return this.hoursOptions.map(option => {
-        return { ...option, disable: this.$moment(option.value, 'HH:mm').isSameOrBefore(this.$moment(this.value.startHour, 'HH:mm')) }
-      });
     },
   },
   methods: {
@@ -129,6 +108,8 @@ export default {
 
   /deep/ .q-field__control
     min-height: 40px
+    .q-field__marginal.q-anchor--skip
+      display: none
 
   .date-item
     @media screen and (min-width: 768px)
