+++ src/components/form/TimeInput.vue
@@ -1,7 +1,7 @@
 <template>
   <div :class="{ 'col-xs-12 col-md-6': !inModal, 'margin-input full-width': inModal }"
     class="input">
-    <div class="row justify-between">
+    <div class="row justify-between" v-if="caption">
       <p :class="['input-caption', { required: requiredField }]">{{ caption }}</p>
       <q-icon v-if="error" name="error_outline" color="secondary" />
     </div>
@@ -10,7 +10,11 @@
       <template v-slot:append>
         <q-icon name="far fa-clock" class="cursor-pointer icon-clock" @click.native="selectTime = !selectTime">
           <q-menu ref="qTimeMenu" anchor="bottom right" self="top right">
-            <q-time :value="value" format24h @input="select" />
+            <q-list dense padding>
+              <q-item v-for="(hour, index) in hoursOptions" :key="index" clickable @click="select(hour.value)">
+                <q-item-section>{{ hour.label }}</q-item-section>
+              </q-item>
+            </q-list>
           </q-menu>
         </q-icon>
       </template>
@@ -19,6 +23,7 @@
 </template>
 
 <script>
+import { PLANNING_VIEW_START_HOUR, PLANNING_VIEW_END_HOUR } from '../../data/constants';
 
 export default {
   name: 'NiTimeInput',
@@ -41,6 +46,27 @@ export default {
       if (this.error) return this.errorLabel;
       return '';
     },
+    hoursOptions () {
+      const range = this.$moment.range(
+        this.$moment().hours(PLANNING_VIEW_START_HOUR).minutes(0),
+        this.$moment().hours(PLANNING_VIEW_END_HOUR).minutes(0)
+      );
+      const hours = Array.from(range.by('hours'));
+      const selectOptions = [];
+      hours.map((hour) => {
+        selectOptions.push({ label: hour.format('HH:mm'), value: hour.format('HH:mm') });
+        if (hour.format('HH') !== `${PLANNING_VIEW_END_HOUR}`) {
+          selectOptions.push({ label: hour.minutes(30).format('HH:mm'), value: hour.minutes(30).format('HH:mm') });
+        }
+      });
+      return selectOptions;
+    },
+    endHourOptions () {
+      return this.hoursOptions.map(option => ({
+        ...option,
+        disable: this.$moment(option.value, 'HH:mm').isSameOrBefore(this.$moment(this.value.startHour, 'HH:mm')),
+      }));
+    },
   },
   methods: {
     select (value) {
@@ -61,4 +87,6 @@ export default {
   .icon-clock
     font-size: 16px
     color: rgba(0,0,0,0.87)
+  .q-list
+    width: 100px
 </style>
